#77 [feat] 랜딩페이지 푸터 모듈화
