+++ src/pages/Index/page.tsx
@@ -1,180 +1,13 @@
-import styled from '@emotion/styled';
-import useWindowSize from '@hooks/useWindowSize';
 import { StyledMain, StyledMainlayout } from '@layout/MainLayout/MainLayout.Style';
-import { LogoImage } from '@layout/SideBar/SideBar.style';
+import IndexFooter from '@components/Index/Footer/Footer';
 import IndexHeader from '@components/Index/Header/Header';
 import IndexFirstSection from '@components/Index/Section/First/FirstSection';
 import IndexLastSection from '@components/Index/Section/Last/LastSection';
 import IndexSecondSection from '@components/Index/Section/Second/SecondSection';
 import IndexThirdSection from '@components/Index/Section/Third/ThirdSection';
-import { media, theme } from '@styles/theme';
-import LogoPNG from '@assets/Logo.png';
-import EnvelopeSVG from '@assets/landing/envelope.svg?react';
-import InstagramSVG from '@assets/landing/instagram.svg?react';
-import LinkedInSVG from '@assets/landing/linkedin.svg?react';
-import PencilSVG from '@assets/landing/pencil.svg?react';
-import ProfileCheerSVG from '@assets/landing/profilecheer.svg?react';
-
-const JuggerButton = styled.button(
-  ({ color, size }: { color: 'primary' | 'secondary'; size: 'small' | 'medium' | 'large' }) => ({
-    ...(color == 'primary'
-      ? {
-          background: theme.color.primary.normal,
-          color: theme.color.label.inverse,
-          ['svg']: {
-            stroke: theme.palette.common[100],
-          },
-        }
-      : {
-          background: theme.palette.coolneutral[98],
-          color: theme.color.label.normal,
-          ['svg']: {
-            stroke: theme.color.label.normal,
-          },
-        }),
-    ...(size == 'small'
-      ? {
-          ...theme.font.body1normal.medium,
-          borderRadius: theme.radius[4],
-          padding: '8px 16px',
-          minWidth: '60px',
-          gap: '4px',
-          ['svg']: {
-            height: '16px',
-            width: '16px',
-          },
-        }
-      : size == 'medium'
-        ? {
-            ...theme.font.headline1.medium,
-            borderRadius: theme.radius[6],
-            padding: '10px 16px',
-            minWidth: '80px',
-            gap: '6px',
-            ['svg']: {
-              height: '20px',
-              width: '20px',
-            },
-          }
-        : {
-            ...theme.font.heading1.medium,
-            borderRadius: theme.radius[8],
-            padding: '12px 24px',
-            minWidth: '120px',
-            gap: '8px',
-            ['svg']: {
-              height: '20px',
-              width: '20px',
-            },
-          }),
-  }),
-  {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    whiteSpace: 'nowrap',
-
-    [':focus']: {
-      outline: 'none',
-    },
-
-    [':disabled']: {
-      background: theme.color.label.disable,
-      color: theme.color.label.assistive,
-      ['svg']: {
-        stroke: theme.color.label.assistive,
-      },
-    },
-  },
-);
-
-const IndexFooterContainer = styled.div({
-  background: theme.palette.blue[20],
-  width: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'start',
-  gap: '36px',
-  padding: '72px',
-  boxSizing: 'border-box',
-
-  ...theme.font.body1normal.medium,
-  color: theme.color.label.inverse,
-
-  [media[480]]: {
-    padding: '32px 20px',
-    gap: '16px',
-
-    ...theme.font.caption1.medium,
-
-    ['img']: {
-      width: '88px',
-      height: '22px',
-    },
-  },
-});
-
-const IndexFooterContent = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '24px',
-
-  [media[480]]: {
-    gap: '12px',
-  },
-});
-
-const IndexFooterButtonContainer = styled.div({
-  display: 'flex',
-  gap: '12px',
-  flexWrap: 'wrap',
-
-  [media[480]]: {
-    gap: '8px',
-  },
-});
-
-const IndexFooterIconContainer = styled.div({
-  display: 'flex',
-  gap: '24px',
-
-  [media[480]]: {
-    gap: '12px',
-  },
-});
-
-const IndexFooter = ({ width }: { width: number }) => {
-  return (
-    <IndexFooterContainer>
-      <LogoImage src={LogoPNG} />
-      <IndexFooterContent>
-        <IndexFooterButtonContainer>
-          <JuggerButton color="secondary" size={width >= 480 ? 'large' : 'small'}>
-            <PencilSVG />
-            의견 남기기
-          </JuggerButton>
-          <JuggerButton color="secondary" size={width >= 480 ? 'large' : 'small'}>
-            <EnvelopeSVG />
-            비즈니스 제안
-          </JuggerButton>
-          <JuggerButton color="secondary" size={width >= 480 ? 'large' : 'small'}>
-            <ProfileCheerSVG />
-            만든이들
-          </JuggerButton>
-        </IndexFooterButtonContainer>
-        <IndexFooterIconContainer>
-          <InstagramSVG />
-          <LinkedInSVG />
-        </IndexFooterIconContainer>
-      </IndexFooterContent>
-      Jugger All rights reserved
-    </IndexFooterContainer>
-  );
-};
+import { theme } from '@styles/theme';
 
 const IndexPage = () => {
-  const width = useWindowSize();
-
   return (
     <StyledMainlayout>
       <StyledMain>
@@ -207,7 +40,7 @@ const IndexPage = () => {
               <IndexLastSection />
             </div>
           </div>
-          <IndexFooter width={width} />
+          <IndexFooter />
         </div>
       </StyledMain>
     </StyledMainlayout>
